+++ Runtime/Serializers/BinarySerializer.cs
@@ -1,9 +1,10 @@
 #if !UNITY_WSA || !UNITY_WINRT
 #define BINARY_AVAILABLE
 #endif
+
+#if BINARY_AVAILABLE
 using System;
 using System.IO;
-#if BINARY_AVAILABLE
 using System.Runtime.Serialization.Formatters.Binary;
 #endif
 
@@ -13,6 +14,7 @@ namespace ActionCode.Persistence
     {
         public string Extension => "bin";
 
+#if BINARY_AVAILABLE
         public string Serialize<T>(T data)
         {
             var formatter = new BinaryFormatter();
@@ -30,6 +32,10 @@ namespace ActionCode.Persistence
             ms.Position = 0;
             return (T)formatter.Deserialize(ms);
         }
+#else
+        public string Serialize<T>(T _) => string.Empty;
+        public T Deserialize<T>(string _) => default;
+#endif
 
         public static bool IsAvailable()
         {